Updated README.md With Template from PurpleBooth
This is a template of a README I have used in the past.  I like this format because it is simple and not specific to the platform you are devleoping for.  I think this is a good step forward for your project to clearly define your problem and make it open source ready